requirements.txt
openai-whisper>=20231117
pydub>=0.25.1
pandas>=2.0.0
python-dateutil>=2.8.2
spacy>=3.7.0
numpy>=1.24.0
reportlab>=4.0.0

